--- code/sys.ui/react.common/src/global.d.mts
@@ -1,4 +1,0 @@
-declare module '*.js?worker';
-declare module '*.js?worker&url';
-declare module '*.js?sharedworker';
-declare module '*.js?sharedworker&url';
