--- code/sys.ui/react.common/src/ui/Event/Event.mts
@@ -1,8 +1,0 @@
-import { useEventsRef as useRef } from './useEventsRef.mjs';
-
-/**
- * Index to [Event] related tools.
- */
-export const Event = {
-  useRef,
-};
