+++ code/sys.ui/react.common/src/ui/Event/Event.ts
@@ -1,0 +1,6 @@
+import { useEventsRef as useRef } from './useEventsRef';
+
+/**
+ * Index to [Event] related tools.
+ */
+export const Event = { useRef } as const;
