+++ code/sys.ui/react.common/src/ui/Event/index.ts
@@ -4,4 +4,4 @@
 
 // export * from '../Event.History';
 // export * from '../Event.List';
-export * from './Event.mjs';
+export * from './Event';
