+++ code/sys.ui/react.common/src/ui/Event/useEventsRef.ts
@@ -1,5 +1,4 @@
 import { useEffect, useRef } from 'react';
-
 import { type t } from '../common';
 
 /**
