+++ code/sys.ui/react.common/src/ui/Focus/Focus.ts
@@ -1,5 +1,5 @@
 import type { RefObject } from 'react';
-import { Wrangle } from './Wrangle';
+import { Wrangle } from './u';
 
 type ElementFocusHandler = (e: FocusEvent) => void;
