+++ code/sys.ui/react.common/src/ui/Icons.ts
