+++ code/sys.ui/react.common/src/ui/PropList/-dev/-common.ts
@@ -1,5 +1,5 @@
 export * from '../common';
-export { Icons } from '../../Icons.mjs';
+export { Icons } from '../../Icons';
 export { PropList } from '..';
 
 export type MyField =
