+++ code/sys.ui/react.common/src/ui/Slider/ui.Ticks.tsx
@@ -1,5 +1,5 @@
-import { Wrangle } from './Wrangle';
 import { Percent, css, type t } from './common';
+import { Wrangle } from './u';
 import { Tick } from './ui.Tick';
 
 export type TicksProps = {
