+++ code/sys.ui/react.common/src/ui/Slider/ui.Track.tsx
@@ -1,5 +1,5 @@
-import { Wrangle } from './Wrangle';
 import { Color, css, type t } from './common';
+import { Wrangle } from './u';
 
 export type TrackProps = {
   enabled: boolean;
