+++ code/sys.ui/react.common/src/ui/Text.Font/Font.ts
@@ -1,5 +1,5 @@
 import { FontContainer as Container } from './ui.FontContainer';
-import { load } from './Font.load.mjs';
+import { load } from './Font.load';
 import { useFont } from './Font.useFont';
 
 /**
