+++ code/sys.ui/react.common/src/ui/Text.Font/index.ts
@@ -1,5 +1,5 @@
 export type { FontCardProps } from './ui.FontCard';
 
-export { Font } from './Font.mjs';
-export { FontCard } from './ui.FontCard';
+export { Font } from './Font';
 export { useFont } from './Font.useFont';
+export { FontCard } from './ui.FontCard';
