+++ code/sys.ui/react.common/src/ui/Text.Keyboard/-dev/-common.ts
@@ -2,6 +2,6 @@ export * from '../common';
 
 export { Card } from '../../Card';
 export { TextSyntax } from '../../Text.Syntax';
-export { Icons } from '../../Icons.mjs';
+export { Icons } from '../../Icons';
 export { Keyboard } from '..';
 export { ObjectView } from '../../ObjectView';
