+++ code/sys.ui/react.common/src/ui/Text/-dev/-SPEC.essay.ts
