+++ code/sys.ui/react.common/src/ui/Text/-dev/-SPEC.tsx
@@ -44,7 +44,7 @@ export default Dev.describe('Text', (e) => {
       });
 
       dev.button('long: Orwell (essay extract - "Why I Write", 1946)', async (e) => {
-        const { Essays } = await import('./DEV.essay.mjs');
+        const { Essays } = await import('./-SPEC.essay');
         const text = Essays.Orwell.extract;
         const words = text.split(' ');
