+++ code/sys.ui/react.common/src/ui/common.ts
@@ -1,3 +1,3 @@
 export * from '../common';
-export { Icons } from './Icons.mjs';
+export { Icons } from './Icons';
 export { Lorem } from '../ui.tools';
